Add method for querying repositories
Adds model classes for deserializing <repository>. Moreover,
add method SUSEStudio.getRepositories() to query the repositories
and tests.

Deserialization of the <repository>.<matches> element with the simple
annotations did not work. I don't think it can be done with
simpleframework, but that would be preferable.
Instead a specific converter Match.ListConverter is used. That converter
needs a different strategy (AnnotationStrategy), which affects the
deserialization of all model classes. This could be an issue.

Signed-off-by: Thomas Haller <thom311@gmail.com>