Let ParserUtils.parseBodyStream throw an exception on error.
Before, ParserUtils.parseBodyStream swallowed parser exceptions
and printed a stacktrace. Such an exception is not at all uncommon
because it can happen when the server sends an unexpected/erroneous
response.

Now any parsing error gets propagated to the caller, because
within the method we cannot properly handle it.

Also adjust all uses of the method, to handle the exception
and rethrow it appropriately.

Signed-off-by: Thomas Haller <thom311@gmail.com>